Tweak setting up the context menus.
