+++ src/background/background.js
@@ -1,4 +1,4 @@
-import './background-setup.js'
 import './open-tutorial.js'
 import './update-content.js'
+import './context-menu.js'
 import '../store/store-background.js'
